requirements_aes.txt
absl-py==2.1.0
accelerate==0.31.0
addict==2.4.0
aeiou
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alias-free-torch==0.0.6
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
argbind==0.3.9
argcomplete==3.1.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
auraloss==0.4.0
backcall==0.2.0
bleach==6.1.0
bokeh==3.1.1
braceexpand==0.1.7
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
clip-anytorch==2.6.0
colorcet==3.1.0
colorlog==6.8.2
comm==0.2.2
configparser==7.0.0
contourpy==1.1.1
cycler==0.12.1
dctorch==0.1.2
debugpy==1.8.5
decorator==5.1.1
descript-audio-codec==1.0.0
descript-audiotools
dnspython==2.6.1
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.7.0
einops-exts==0.0.4
ema-pytorch==0.2.3
email_validator==2.1.2
encodec==0.1.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastcore==1.5.46
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.15.1
fire==0.6.0
flatten-dict==0.4.2
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
ftfy==6.2.0
future==1.0.0
gin-config==0.5.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.30.0
google-auth-oauthlib==1.0.0
gradio==4.36.1
gradio_client==1.0.1
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
holoviews==1.17.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
Hydra==2.5
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.7
imageio==2.34.1
importlib-resources==5.12.0
importlib_metadata==7.1.0
inquirerpy==0.3.4
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonmerge==1.9.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
julius==0.2.7
jupyter_client==8.6.2
jupyter_core==5.7.2
k-diffusion==0.1.1
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
laion-clap
lazy_loader==0.4
librosa
lightning-utilities==0.11.2
linkify-it-py==2.0.3
llvmlite==0.41.1
local-attention==1.8.6
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
mmcv==2.2.0
mmengine==0.10.4
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mutagen==1.47.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.1
noisereduce==3.0.2
numba==0.58.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.10.0.84
orjson==3.10.5
packaging==24.1
pandas==2.0.2
panel==1.2.3
param==2.1.0
parso==0.8.4
pathtools==0.1.2
pedalboard==0.7.4
pexpect==4.9.0
pfzy==0.3.4
pickleshare==0.7.5
pillow==10.3.0
pipx==1.2.1
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
plotly==5.22.0
pooch==1.8.2
prefigure==0.0.9
progressbar==2.5
prompt_toolkit==3.0.47
protobuf==3.19.6
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
pyloudnorm==0.1.1
pynndescent==0.5.13
pyparsing==3.1.2
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.1.0
pytz==2024.1
pyviz_comms==3.0.2
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.1.0
randomname==0.2.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
resampy==0.4.3
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruff==0.4.9
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.5.1
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile
soxr==0.3.7
stable-audio-tools
stack-data==0.6.3
starlette==0.37.2
sympy==1.12.1
tenacity==8.4.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch
torch-stoi==0.2.1
torchaudio==2.2.1
torchdiffeq==0.2.4
torchlibrosa==0.1.0
torchmetrics==0.11.4
torchsde==0.2.6
torchvision==0.17.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
trampoline==0.1.2
transformers==4.41.2
triton==2.2.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
umap-learn==0.5.6
urllib3==2.2.2
userpath==1.9.1
uvicorn==0.30.1
uvloop==0.19.0
v-diffusion-pytorch==0.0.2
vector-quantize-pytorch==1.9.14
wandb==0.15.4
watchfiles==0.22.0
wcwidth==0.2.13
webdataset==0.2.48
webencodings==0.5.1
websockets
Werkzeug==3.0.3
wget==3.2
x-transformers==1.26.6
xyzservices==2024.6.0
yapf==0.40.2
yarl==1.9.4
yt-dlp 
zipp==3.19.2